backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database and Backend-as-a-Service
supabase>=2.0.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Google OAuth and API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# Google Gemini AI
google-generativeai>=0.3.0

# Email service
sendgrid>=6.10.0

# Data validation
pydantic>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.1.0

# HTTP requests
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0

# Rate limiting
slowapi>=0.1.9
